refactor: JPA Auditing 설정과 AuditorAware 설정 통합
